+++ packages/api-reference/src/features/Operation/layouts/ClassicLayout.vue
@@ -114,13 +114,20 @@ const { copyToClipboard } = useClipboard()
       <XBadges
         :badges="operation['x-badges']"
         position="after" />
-      <TestRequestButton
-        v-if="active && !isWebhook"
-        :method="method"
-        :path="path" />
-      <ScalarIconPlay
-        v-else-if="!config?.hideTestRequestButton"
-        class="endpoint-try-hint size-4.5" />
+      <template v-if="!config?.hideTestRequestButton">
+        <TestRequestButton
+          v-if="active && !isWebhook"
+          :method="method"
+          :path="path" />
+        <ScalarIconPlay
+          v-else
+          class="endpoint-try-hint size-4.5" />
+      </template>
+      <span
+        v-if="config.showOperationId && operation.operationId"
+        class="font-code text-sm">
+        {{ operation.operationId }}
+      </span>
       <ScalarIconButton
         class="endpoint-copy p-0.5"
         :icon="ScalarIconCopy"
@@ -183,10 +190,10 @@ const { copyToClipboard } = useClipboard()
             class="operation-example-card"
             :clientOptions="clientOptions"
             fallback
+            :isWebhook="isWebhook"
             :method="method"
             :operation="operation"
             :path="path"
-            :isWebhook="isWebhook"
             :securitySchemes="securitySchemes"
             :selectedClient="store.workspace['x-scalar-default-client']"
             :selectedServer="server" />
@@ -435,8 +442,8 @@ const { copyToClipboard } = useClipboard()
 
 .operation-example-card {
   position: sticky;
-  top: calc(var(--refs-header-height) + 24px);
-  max-height: calc(((var(--full-height) - var(--refs-header-height)) - 48px));
+  top: calc(var(--refs-viewport-offset) + 24px);
+  max-height: calc(var(--refs-viewport-height) - 48px);
 }
 
 @media (max-width: 600px) {