+++ packages/components/src/components/ScalarCheckboxInput/index.ts
@@ -1,0 +1,6 @@
+export { default as ScalarCheckboxInput } from './ScalarCheckboxInput.vue'
+export { default as ScalarCheckboxGroup } from './ScalarCheckboxGroup.vue'
+export { default as ScalarCheckboxRadioGroup } from './ScalarCheckboxRadioGroup.vue'
+export { default as ScalarCheckbox } from './ScalarCheckbox.vue'
+
+export type { ScalarCheckboxOption, ScalarCheckboxType } from './types'