+++ packages/components/src/components/ScalarTextInput/ScalarTextInput.vue
@@ -13,6 +13,12 @@ export default {}
 import { useBindCx } from '@scalar/use-hooks/useBindCx'
 import { onMounted, ref } from 'vue'
 
+import { ScalarFormInput } from '../ScalarForm'
+
+const { readonly } = defineProps<{
+  readonly?: boolean
+}>()
+
 const model = defineModel<string>()
 
 const input = ref<HTMLInputElement>()
@@ -24,15 +30,22 @@ onMounted(() => {
   // Force autofocus if the input has the autofocus attribute
   if ('autofocus' in otherAttrs.value) input.value?.focus()
 })
+
+function handleClick() {
+  if (readonly) {
+    input.value?.select() // If readonly, select the input on click
+  } else {
+    input.value?.focus() // If not readonly, focus the input on click
+  }
+}
 </script>
 <template>
-  <div
+  <ScalarFormInput
+    is="div"
     v-bind="
-      classCx(
-        'bg-b-1.5 flex cursor-text items-center gap-0.75 rounded-md border px-3 py-2.5 outline-offset-[-1px] focus-within:bg-b-1 has-[input:focus-visible]:outline',
-      )
+      classCx('cursor-text text-c-1', readonly ? '' : 'focus-within:bg-b-1')
     "
-    @click="input?.focus()">
+    @click="handleClick">
     <div class="flex flex-1 relative">
       <span
         v-if="$slots.prefix"
@@ -42,6 +55,8 @@ onMounted(() => {
       <input
         ref="input"
         v-model="model"
+        :readonly
+        :aria-readonly="readonly || undefined"
         class="z-1 min-w-0 flex-1 border-none bg-transparent text-sm placeholder:font-[inherit] focus-within:outline-none"
         v-bind="otherAttrs" />
       <div
@@ -61,5 +76,5 @@ onMounted(() => {
       </div>
     </div>
     <slot name="aside" />
-  </div>
+  </ScalarFormInput>
 </template>